% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create \code{filename} from the \code{year} provided.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year}
}
\value{
a string with the following structure: accident_YEAR.csv.bz2
}
\description{
Function that takes a year input and outputs a string with the following structure: accident_YEAR.csv.bz2
Made to retrieve data  from the \href{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}{US National Highway Traffic Safety Administration's Fatality Analysis Reporting System}
}
\examples{
\dontrun{
filepath <- make_filename(2012)
}
}
